Added User ID underneath Member Name
