+++ Android App/ColonelFund/app/src/main/res/layout/event_list_item.xml
@@ -69,7 +69,7 @@
         android:textSize="18dp" />
 
     <TextView
-        android:id="@+id/eventUser"
+        android:id="@+id/userID"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:layout_alignTop="@id/memberName"
@@ -94,7 +94,7 @@
         android:layout_alignParentLeft="true"
         android:layout_alignParentRight="true"
         android:layout_alignParentStart="true"
-        android:layout_below="@+id/eventUser"
+        android:layout_below="@+id/userID"
         android:progress="45" />
 
 </RelativeLayout>
