+++ Android App/ColonelFund/app/src/main/res/layout/member_list_item.xml
@@ -25,12 +25,27 @@
         android:id="@+id/memberName"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_alignBaseline="@+id/member_initials"
-        android:layout_alignParentRight="true"
-        android:layout_marginRight="0dp"
-        android:layout_toEndOf="@+id/member_initials"
         android:layout_marginLeft="8dp"
+        android:layout_marginTop="10dp"
+        android:layout_toEndOf="@+id/member_initials"
         android:layout_toRightOf="@+id/member_initials"
+        android:gravity="left"
+        android:text="Member Name"
         android:textSize="18dp" />
+
+    <TextView
+        android:id="@+id/userID"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:layout_alignLeft="@id/memberName"
+        android:layout_alignRight="@id/memberName"
+        android:layout_alignTop="@id/memberName"
+        android:layout_below="@+id/memberName"
+        android:layout_marginLeft="16dp"
+        android:layout_marginRight="0dp"
+        android:layout_marginTop="18dp"
+        android:gravity="left"
+        android:text="User ID"
+        android:textSize="14dp" />
 
 </RelativeLayout>
